# showcase/openapi.yaml
openapi: 3.1.0
info:
  title: TEOB-TS Showcase API
  description: |
    Full-stack showcase for the TEOB-TS event-sourcing framework.
    Demonstrates pure aggregates (Gift Card, Todo), Petri Net flows (Signup),
    and AI integration (chat with entity-aware tools and knowledge search).
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/chat/sessions:
    post:
      tags: [AI Chat]
      summary: Create a new chat session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                systemPrompt:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
  /api/chat/sessions/{id}/messages:
    post:
      tags: [AI Chat]
      summary: Send a message and get AI response (may use tools)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  toolCalls:
                    type: array
                    nullable: true
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        result: {}
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: AI not configured

    get:
      tags: [AI Chat]
      summary: Get conversation history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                          enum: [user, assistant]
                        content:
                          type: string
  /api/chat/sessions/{id}:
    delete:
      tags: [AI Chat]
      summary: Delete a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '404':
          $ref: '#/components/responses/NotFound'

  # --- System ---
...
components:
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [ok]
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string